feat: setting baseurl via environment variable and improving detection of internal requests
