+++ src/runtime/composables/local/useAuthState.ts
@@ -22,7 +22,7 @@ export interface UseAuthStateReturn extends CommonUseAuthStateReturn<SessionData
   setToken: (newToken: string | null) => void
   clearToken: () => void
   _internal: {
-    baseURL: string
+    origin?: string
     pathname: string
     rawTokenCookie: CookieRef<string | null>
   }
