+++ src/runtime/helpers.ts
@@ -2,12 +2,25 @@
 import { parseURL } from 'ufo'
 import type { DeepRequired } from 'ts-essentials'
 import type { ProviderAuthjs, ProviderLocal, SupportedAuthProviders } from './types'
-import type { useRuntimeConfig } from '#imports'
+import { extractFromRuntimeConfig } from './utils/extractFromRuntimeConfig'
+import { useRuntimeConfig } from '#imports'
 
 export const isProduction = process.env.NODE_ENV === 'production'
 
-export function getOriginAndPathnameFromURL(url: string) {
-  const { protocol, host, pathname } = parseURL(url)
+export function getOriginAndPathnameFromURL(url: string, envOriginKey?: string) {
+  let baseUrl = url
+  const config = useRuntimeConfig()
+
+  // Prioritize environment variable if set
+  if (envOriginKey) {
+    const envFromRuntimeConfig = extractFromRuntimeConfig(config, envOriginKey)
+    const envOrigin = envFromRuntimeConfig ?? process.env[envOriginKey]
+    if (envOrigin) {
+      baseUrl = envOrigin
+    }
+  }
+
+  const { protocol, host, pathname } = parseURL(baseUrl)
 
   let origin
   if (host && protocol) {
